+++ dzVents/runtime/tests/data/__data_global_data.lua
@@ -1,0 +1,9 @@
+-- Persistent Data
+local multiRefObjects = {
+
+} -- multiRefObjects
+local obj1 = {
+	["g"] = 666;
+	["h"] = true;
+}
+return obj1