+++ dzVents/documentation/history.md
@@ -11,6 +11,7 @@
  - Added the lua Lodash library (http://axmat.github.io/lodash.lua, MIT license).
  - Fixed documentation about levelNames for selector switches and added the missing levelName.
  - Added silent argument (boolean) to device.switchOn() and device.switchOff() commands. When set to true, no follow-up events are triggered switching the device. Using switchOn(true) will not allow you to set timing options!
+ - Moved dzVents runtime code away from the /path/to/domoticz/scripts/dzVents folder as this scripts folder contains user stuff.
 
 [2.2.0]